+++ sei-ibc-go/modules/light-clients/06-solomachine/types/client_state.go
@@ -22,7 +22,7 @@ var _ exported.ClientState = (*ClientState)(nil)
 func NewClientState(latestSequence uint64, consensusState *ConsensusState, allowUpdateAfterProposal bool) *ClientState {
 	return &ClientState{
 		Sequence:                 latestSequence,
-		FrozenSequence:           0,
+		IsFrozen:                 false,
 		ConsensusState:           consensusState,
 		AllowUpdateAfterProposal: allowUpdateAfterProposal,
 	}
@@ -45,16 +45,11 @@ func (cs ClientState) GetLatestHeight() exported.Height {
 // - Active: if frozen sequence is 0
 // - Frozen: otherwise solo machine is frozen
 func (cs ClientState) Status(_ sdk.Context, _ sdk.KVStore, _ codec.BinaryCodec) exported.Status {
-	if cs.FrozenSequence != 0 {
+	if cs.IsFrozen {
 		return exported.Frozen
 	}
 
 	return exported.Active
-}
-
-// IsFrozen returns true if the client is frozen.
-func (cs ClientState) IsFrozen() bool {
-	return cs.FrozenSequence != 0
 }
 
 // GetProofSpecs returns nil proof specs since client state verification uses signatures.
@@ -438,10 +433,6 @@ func produceVerificationArgs(
 	}
 	// sequence is encoded in the revision height of height struct
 	sequence := height.GetRevisionHeight()
-	if cs.IsFrozen() {
-		return nil, nil, 0, 0, clienttypes.ErrClientFrozen
-	}
-
 	if prefix == nil {
 		return nil, nil, 0, 0, sdkerrors.Wrap(commitmenttypes.ErrInvalidPrefix, "prefix cannot be empty")
 	}