+++ sei-ibc-go/modules/light-clients/06-solomachine/types/misbehaviour_handle_test.go
@@ -38,16 +38,6 @@ func (suite *SoloMachineTestSuite) TestCheckMisbehaviourAndUpdateState() {
 				}, true,
 			},
 			{
-				"client is frozen",
-				func() {
-					cs := solomachine.ClientState()
-					cs.FrozenSequence = 1
-					clientState = cs
-					misbehaviour = solomachine.CreateMisbehaviour()
-				},
-				false,
-			},
-			{
 				"wrong client state type",
 				func() {
 					clientState = &ibctmtypes.ClientState{}
@@ -264,7 +254,7 @@ func (suite *SoloMachineTestSuite) TestCheckMisbehaviourAndUpdateState() {
 
 				if tc.expPass {
 					suite.Require().NoError(err)
-					suite.Require().True(clientState.(*types.ClientState).FrozenSequence != 0, "client not frozen")
+					suite.Require().True(clientState.(*types.ClientState).IsFrozen, "client not frozen")
 				} else {
 					suite.Require().Error(err)
 					suite.Require().Nil(clientState)