+++ sei-ibc-go/modules/light-clients/06-solomachine/types/proposal_handle_test.go
@@ -77,7 +77,7 @@ func (suite *SoloMachineTestSuite) TestCheckSubstituteAndUpdateState() {
 
 					suite.Require().Equal(substituteClientState.(*types.ClientState).ConsensusState, updatedClient.(*types.ClientState).ConsensusState)
 					suite.Require().Equal(substituteClientState.(*types.ClientState).Sequence, updatedClient.(*types.ClientState).Sequence)
-					suite.Require().Equal(uint64(0), updatedClient.(*types.ClientState).FrozenSequence)
+					suite.Require().Equal(false, updatedClient.(*types.ClientState).IsFrozen)
 				} else {
 					suite.Require().Error(err)
 					suite.Require().Nil(updatedClient)